+++ tests/Db/QueryBuilder/Condition/Builder/BetweenColumnsBuilderTest.php
@@ -20,14 +20,13 @@ final class BetweenColumnsBuilderTest extends TestCase
     {
         $db = $this->getConnection();
 
-        $betweenColumnsCondition = new BetweenColumns(42, 'BETWEEN', '1', '100');
+        $betweenColumnsCondition = new BetweenColumns(42, '1', '100');
         $params = [];
 
         $this->assertSame(
-            ':qp0 BETWEEN [1] AND [100]',
+            '42 BETWEEN [1] AND [100]',
             (new BetweenColumnsBuilder($db->getQueryBuilder()))->build($betweenColumnsCondition, $params)
         );
-
-        $this->assertEquals([':qp0' => 42], $params);
+        $this->assertSame([], $params);
     }
 }